--- weed/s3api/s3lifecycle/dispatcher/dispatch_ticks_default.go
@@ -1,10 +1,0 @@
-//go:build !s3tests
-
-package dispatcher
-
-import "time"
-
-const (
-	defaultDispatchTick   = 5 * time.Second
-	defaultCheckpointTick = 30 * time.Second
-)
