--- weed/s3api/s3lifecycle/scheduler/refresh_default.go
@@ -1,7 +1,0 @@
-//go:build !s3tests
-
-package scheduler
-
-import "time"
-
-const defaultRefreshInterval = 5 * time.Minute
